Add photo

+++ src/de/raphaelmichel/lendlist/storage/Database.java
@@ -8,19 +8,16 @@ import android.util.Log;
 public class Database extends SQLiteOpenHelper {
 
 	public static final String DATABASE_NAME = "objects.db";
-	public static final int DATABASE_VERSION = 6; // REPLACE ONUPGRADE IF YOU
+	public static final int DATABASE_VERSION = 7; // REPLACE ONUPGRADE IF YOU
 													// CHANGE THIS
 
-	private static final String DATABASE_CREATE = "create table "
-			+ "objects ( id integer primary key autoincrement,"
-			+ " direction text," + " thing text," + " person text,"
-			+ " contact_id integer," + " until integer," + " date integer,"
-			+ " returned integer," + " contact_lookup text" + ");" 
-			+ "create table photos ( id integer primary key autoincrement,"
-			+ "object integer," + " uri text"+ ");";
 	public static final String[] COLUMNS = { "id AS _id", "direction", "thing",
 			"person", "contact_id", "until", "date", "returned",
 			"contact_lookup" };
+
+	public static final String[] COLUMNS_PHOTOS = { "id AS _id", "object",
+			"uri" };
+
 	public static final String OBJECT_TABLE = "objects";
 	public static final String OBJECT_WHERE_ID = "id = ?";
 	public static final String PHOTO_TABLE = "photos";
@@ -29,16 +26,29 @@ public class Database extends SQLiteOpenHelper {
 
 	@Override
 	public void onCreate(SQLiteDatabase db) {
-		db.execSQL(DATABASE_CREATE);
+		db.execSQL("create table "
+				+ "objects ( id integer primary key autoincrement,"
+				+ " direction text," + " thing text," + " person text,"
+				+ " contact_id integer," + " until integer," + " date integer,"
+				+ " returned integer," + " contact_lookup text" + ");");
+		db.execSQL("create table photos ( id integer primary key autoincrement,"
+				+ "object integer," + " uri text" + ");");
 	}
 
 	@Override
 	public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-		Log.w(Database.class.getName(), "Upgrading database from version "
-				+ oldVersion + " to " + newVersion
-				+ ", which will destroy all old data");
-		db.execSQL("DROP TABLE IF EXISTS objects");
-		onCreate(db);
+		if(db.isReadOnly()) db = getWritableDatabase();
+		if (oldVersion == 6 && newVersion == 7) {
+			db.execSQL("create table photos ( id integer primary key autoincrement,"
+					+ "object integer," + " uri text" + ");");
+		} else {
+			Log.w(Database.class.getName(), "Upgrading database from version "
+					+ oldVersion + " to " + newVersion
+					+ ", which will destroy all old data");
+			db.execSQL("DROP TABLE IF EXISTS objects");
+			db.execSQL("DROP TABLE IF EXISTS photos");
+			onCreate(db);
+		}
 	}
 
 	public Database(Context context) {
